+++ PiPiPaPaBoom/Assets/Scripts/Enemy/Cucumber.cs
@@ -13,7 +13,7 @@ public class Cucumber : Enemy, IDamageable
     }
 
 
-    public void GetHit(float damage)
+    public void GetHit(int damage)
     {
         currentHealth -= damage;
         if (currentHealth < 1) {
