+++ PiPiPaPaBoom/Assets/Scripts/IDamageable.cs
@@ -1,5 +1,5 @@
 
 public interface IDamageable
 {
-    void GetHit(float damage);
+    void GetHit(int damage);
 }
